Fix relative URL root and tests
Relative URL root had been set ad environment variable because it
broke tests when set. However, the issue is that it had been set with
forward slash at the beginning which causes different behavior between
test and the rest of the application.

This stops relying on environment variables which better for such
exploratory project.